+++ spec/models/calagator/event_search_spec.rb
@@ -114,37 +114,5 @@ module Calagator
         expect(Event::SearchEngine).not_to be_score
       end
     end
-
-    describe "Sunspot" do
-      around do |example|
-        server_running = begin
-          # Try opening the configured port. If it works, it's running.
-          TCPSocket.new("127.0.0.1", Sunspot::Rails.configuration.port).close
-          true
-        rescue Errno::ECONNREFUSED
-          false
-        end
-
-        if server_running
-          Event::SearchEngine.use(:sunspot)
-          Venue::SearchEngine.use(:sunspot)
-          described_class.reindex
-          Venue.reindex
-          example.run
-        else
-          pending "Solr not running. Start with `rake sunspot:solr:start RAILS_ENV=test`"
-        end
-      end
-
-      it_behaves_like "#search"
-
-      it "is using the sunspot search engine" do
-        expect(Event::SearchEngine.kind).to eq(:sunspot)
-      end
-
-      it "provides a score" do
-        expect(Event::SearchEngine).to be_score
-      end
-    end
   end
 end